requirements.txt
aiohttp
aiosqlite
pyyaml
pytest-aiohttp
